refactor: Create classes/ in annotations
